Narrow token cache scope to the provider level
Previously, we cached tokens in a global map. When any token read or
write happened, we took an exclusive lock.

On a cache miss, we make network calls to get a valid token. If those
network calls block, all requests that try to query will block, for all
tenants sharing the same pod.

This change scopes the cache to the provider level.